设备开关联动

+++ src/main/java/com/radebit/smartpi/service/impl/DeviceServiceImpl.java
@@ -2,7 +2,9 @@ package com.radebit.smartpi.service.impl;
 
 import com.radebit.smartpi.mapper.DeviceMapper;
 import com.radebit.smartpi.model.po.Device;
+import com.radebit.smartpi.model.po.DeviceSwitch;
 import com.radebit.smartpi.service.DeviceService;
+import com.radebit.smartpi.service.DeviceSwitchService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
@@ -18,6 +20,9 @@ public class DeviceServiceImpl implements DeviceService {
 
     @Autowired
     private DeviceMapper deviceMapper;
+
+    @Autowired
+    private DeviceSwitchService deviceSwitchService;
 
     @Override
     public List<Device> findAll() {
@@ -51,11 +56,35 @@ public class DeviceServiceImpl implements DeviceService {
 
     @Override
     public Integer delete(int id) {
+        if (deviceSwitchService.delete(id)!=1){
+            return 0;
+        }
         return deviceMapper.delete(id);
     }
 
     @Override
     public Integer add(Device device) {
-        return deviceMapper.add(device);
+        Integer add = deviceMapper.add(device);
+
+        DeviceSwitch deviceSwitch = new DeviceSwitch();
+        deviceSwitch.setDeviceId(device.getId());
+        deviceSwitch.setWater1(0);
+        deviceSwitch.setWater2(0);
+        deviceSwitch.setWater3(0);
+        deviceSwitch.setHeat(0);
+        deviceSwitch.setFan(0);
+        deviceSwitch.setHumidifier(0);
+        deviceSwitch.setRed(0);
+        deviceSwitch.setBlue(0);
+        deviceSwitch.setGreen(0);
+        deviceSwitch.setYellow(0);
+        deviceSwitch.setIllumination(0);
+        deviceSwitch.setUltraviolet(0);
+
+        if (deviceSwitchService.add(deviceSwitch) != 1) {
+            return 0;
+        }
+
+        return add;
     }
 }
